Resolve issue #25
The issue was that, on the documentation page, there will sometimes
be extra information in addition to the string representing the
Type. In this case, the InstanceType contained information about a
"list of valid instance types," and so "list" was being picked up.
Now, we simply ignore blocks of HTML with commas in them, because
type information is likely in the next set of tags.